Major refactor in archive.py
Removed setter methods for `start_date` and `end_date` attributes in the
`WeatherArchive` class and replaced them with the `set_start_date` and
`set_end_date` methods for following a more static typing approach as the
type for the arguments and variable couldn't be matched in the setter methods.
Also updated the `__init__` method in correspondance with the changes.