Update `BaseForecast` class in base.py
Updated the `__init__` method of the `BaseForecast` and added
`timeout` as a parameter in the method to accept the maximum wait
duration for a response from the API endpoint. Also added it as
an argument when invoking the `__init__` method of the parent class
and updated `__slots__` with the same.