Refactor a method in BaseMeteor class in base.py
Refactor the timeout.setter method to explicitly verify the timeout
value rather than calling the `verify_timeout` function as the function
was removed in a previous commit. This change was made as this method
was the sole user of the function and the mechanism was not used anywhere
else in the codebase.